+++ RedPandaIDE/settingsdialog/executorproblemsetwidget.h
@@ -21,6 +21,8 @@ private:
 protected:
     void doLoad() override;
     void doSave() override;
+private slots:
+    void on_chkOnlyMonospaced_stateChanged(int arg1);
 };
 
 #endif // EXECUTORPROBLEMSETWIDGET_H